+++ sdk/core/azure-core/src/http/curl/curl.cpp
@@ -150,6 +150,22 @@ void static inline SetHeader(Azure::Core::Http::RawResponse& response, std::stri
       reinterpret_cast<uint8_t const*>(header.data() + header.size()));
 }
 
+static inline std::string GetHeadersAsString(Azure::Core::Http::Request const& request)
+{
+  std::string requestHeaderString;
+
+  for (auto const& header : request.GetHeaders())
+  {
+    requestHeaderString += header.first; // string (key)
+    requestHeaderString += ": ";
+    requestHeaderString += header.second; // string's value
+    requestHeaderString += "\r\n";
+  }
+  requestHeaderString += "\r\n";
+
+  return requestHeaderString;
+}
+
 // Writes an HTTP request with RFC 7230 without the body (head line and headers)
 // https://tools.ietf.org/html/rfc7230#section-3.1.1
 static inline std::string GetHTTPMessagePreBody(Azure::Core::Http::Request const& request)
@@ -160,7 +176,7 @@ static inline std::string GetHTTPMessagePreBody(Azure::Core::Http::Request const
   httpRequest += " /" + url + " HTTP/1.1\r\n";
 
   // headers
-  httpRequest += request.GetHeadersAsString();
+  httpRequest += GetHeadersAsString(request);
 
   return httpRequest;
 }